Enum dumbserve::errors::ServiceError

pub enum ServiceError {

    Unauthorized,
    InternalServerError,
    ClosedForRegistration,
    NotAnEmail,
    NotAUrl,
    WrongPassword,
    ProfainityError,
    BlacklistError,
    UsernameCaseMappedError,
    PasswordTooShort,
    PasswordTooLong,
    PasswordsDontMatch,
    UsernameTaken,
    EmailTaken,

}

Variants

Unauthorized

InternalServerError

ClosedForRegistration

NotAnEmail

NotAUrl

WrongPassword

ProfainityError

when the value passed contains profainity

BlacklistError

when the value passed contains blacklisted words see blacklist

UsernameCaseMappedError

when the value passed contains characters not present in UsernameCaseMapped profile

PasswordTooShort

PasswordTooLong

PasswordsDontMatch

UsernameTaken

when the a username is already taken

EmailTaken

email is already taken